Complicated conflicts below in the rhnsat sql files.

Conflicts:
	java/spacewalk-java.spec
	schema/spacewalk/rhnsat/tables/Makefile.deps
	schema/spacewalk/rhnsat/tables/rhnActionKickstart.sql
	schema/spacewalk/rhnsat/tables/rhnErrata.sql
	schema/spacewalk/rhnsat/tables/rhnErrataNotificationQueue.sql
	schema/spacewalk/rhnsat/tables/rhnErrataQueue.sql
	schema/spacewalk/rhnsat/tables/rhnKickstartScript.sql
	schema/spacewalk/rhnsat/tables/rhnServerNeededCache_indexes.sql
	schema/spacewalk/rhnsat/tables/rhn_command_parameter_data.sql
	schema/spacewalk/rhnsat/tables/time_series.sql
	schema/spacewalk/rhnsat/views/rhnPrivateErrataMail.sql
	spacewalk/setup/lib/Spacewalk/Setup.pm

+++ proxy/installer/configure-proxy.sh
@@ -20,6 +20,56 @@ options:
             For use only with --answer-file. If the --answer-file doesn't
             provide a required response, default answer is used.
 
+  --version=VERSION
+			Version of Spacewalk Proxy Server you want to activate.
+  --rhn-parent=RHN_PARENT
+			Your parent Spacewalk server.
+  --traceback-email=TRACEBACK_EMAIL
+			Email to which tracebacks should be sent.
+  --use-ssl=USE_SSL
+			1  if  Spacewalk  Proxy Server should communicate with parent over SSL. 
+			0 otherwise. Even if disabled, client can still use SSL to connect 
+			to Spacewalk Proxy Server.
+  --ca-chain=CA_CHAIN
+			The CA cert used to verify the ssl connection to parent.
+  --http-proxy=HTTP_PROXY
+			HTTP proxy in host:port format, e.g. squid.redhat.com:3128
+  --http-username=HTTP_USERNAME
+			The username for an authenticated proxy.
+  --http-password=HTTP_PASSWORD
+			The password to use for an authenticated proxy.
+  --ssl-org=SSL_ORG
+			Organization name to be used in SSL certificate.
+  --ssl-orgunit=SSL_ORGUNIT
+			Organization unit name to be used in SSL certificate.
+  --ssl-common=SSL_COMMON
+			Common name to be used in SSL certificate.
+  --ssl-city=SSL_CITY
+			City to be used in SSL certificate.
+  --ssl-state=SSL_STATE
+			State to be used in SSL certificate.
+  --ssl-country=SSL_COUNTRY
+			Two letters country code to be used in SSL certificate.
+  --ssl-email=SSL_EMAIL
+			Email to be used in SSL certificate.
+  --ssl-password=SSL_PASSWORD
+			Password to be used for SSL CA certificate.
+  --install-monitoring=Y
+			Y if monitoring should be installed. Any other value means that 
+			monitoring will not be installed.
+  --enable-scout=1
+			1 to enable monitoring scout, 0 otherwise.
+  --monitoring-parent=MONITORING_PARENT
+			Name of the parent for your scout. Usually the same value as in 
+			RHN_PARENT.
+  --monitoring-parent-ip=MONITORING_PARENT_IP
+			IP address of MONITORING_PARENT
+  --scout-shared-key=SCOUT_SHARED_KEY
+			Your scout shared key (can be found on the parent in 
+			/etc/rhn/cluster.ini as key scoutsharedkey).
+  --populate-config-channel=Y
+			Y if config chanel should be created and configuration files in that channel
+			updated. Configuration channel will be named rhn_proxy_config_\${SYSTEM_ID}.		
 HELP
 	exit
 }
@@ -29,8 +79,32 @@ INTERACTIVE=1
 while [ $# -ge 1 ]; do
 	case $1 in
             --help | -h)  print_help;;
-            --answer-file=*) . $(echo $1 | cut -d= -f2);;
+            --answer-file=*) . $(echo $1 | cut -d= -f2-);;
             --non-interactive) INTERACTIVE=0;;
+
+			--version=*) VERSION=$(echo $1 | cut -d= -f2-);;
+			--rhn-parent=*) RHN_PARENT=$(echo $1 | cut -d= -f2-);;
+			--traceback-email=*) TRACEBACK_EMAIL=$(echo $1 | cut -d= -f2-);;
+			--use-ssl=*) USE_SSL=$(echo $1 | cut -d= -f2-);;
+			--ca-chain=*) CA_CHAIN=$(echo $1 | cut -d= -f2-);;
+			--http-proxy=*) HTTP_PROXY=$(echo $1 | cut -d= -f2-);;
+			--http-username=*) HTTP_USERNAME=$(echo $1 | cut -d= -f2-);;
+			--http-password=*) HTTP_PASSWORD=$(echo $1 | cut -d= -f2-);;
+			--ssl-org=*) SSL_ORG=$(echo $1 | cut -d= -f2-);;
+			--ssl-orgunit=*) SSL_ORGUNIT=$(echo $1 | cut -d= -f2-);;
+			--ssl-common=*) SSL_COMMON=$(echo $1 | cut -d= -f2-);;
+			--ssl-city=*) SSL_CITY=$(echo $1 | cut -d= -f2-);;
+			--ssl-state=*) SSL_STATE=$(echo $1 | cut -d= -f2-);;
+			--ssl-country=*) SSL_COUNTRY=$(echo $1 | cut -d= -f2-);;
+			--ssl-email=*) SSL_EMAIL=$(echo $1 | cut -d= -f2-);;
+			--ssl-password=*) SSL_PASSWORD=$(echo $1 | cut -d= -f2-);;
+			--install-monitoring=*) INSTALL_MONITORING=$(echo $1 | cut -d= -f2-);;
+			--enable-scout=*) ENABLE_SCOUT=$(echo $1 | cut -d= -f2-);;
+			--monitoring-parent=*) MONITORING_PARENT_IP=$(echo $1 | cut -d= -f2-);;
+			--monitoring-parent-ip=*) MONITORING_PARENT_IP=$(echo $1 | cut -d= -f2-);;
+			--scout-shared-key=*) SCOUT_SHARED_KEY=$(echo $1 | cut -d= -f2-);;
+			--populate-config-channel=*) POPULATE_CONFIG_CHANNEL=$(echo $1 | cut -d= -f2-);;
+			*) echo Error: Invalid option $1
     esac
     shift
 done
@@ -103,6 +177,11 @@ HTMLPUB_DIR=/var/www/html/pub
 JABBERD_DIR=/etc/jabberd
 SQUID_DIR=/etc/squid
 
+if [ ! -r $SYSCONFIG_DIR/systemid ]; then
+	echo ERROR: RHN Proxy does not appear to be registered
+	exit 2
+fi
+
 SYSTEM_ID=$(/usr/bin/xsltproc /usr/share/rhn/get_system_id.xslt $SYSCONFIG_DIR/systemid | cut -d- -f2)
 
 DIR=/usr/share/doc/proxy/conf-template
@@ -126,6 +205,15 @@ default_or_input "Use SSL" USE_SSL 'Y/n'
 USE_SSL=$(yes_no $USE_SSL)
 
 default_or_input "CA Chain" CA_CHAIN $(awk -F'[=;]' '/sslCACert=/ {a=$2} END { print a}' $SYSCONFIG_DIR/up2date)
+
+if [ "$RHN_PARENT" != "xmlrpc.rhn.redhat.com" -a ! -f /root/ssl-build/RHN-ORG-PRIVATE-SSL-KEY ] && ! diff $CA_CHAIN /root/ssl-build/RHN-ORG-TRUSTED-SSL-KEY &>/dev/null; then
+	cat <<CA_KEYS
+Please do copy your CA key and public certificate from $RHN_PARENT to 
+/root/ssl-build directory. You may want to execute this command:
+ scp 'root@$RHN_PARENT:/root/ssl-build/RHN-ORG-{PRIVATE-SSL-KEY,TRUSTED-SSL-CERT}' /root/ssl-build
+CA_KEYS
+	exit 1
+fi
 
 if ! /sbin/runuser nobody -s /bin/sh --command="[ -r $CA_CHAIN ]" ; then
 	echo Error: File $CA_CHAIN is not readable by nobody user.
@@ -184,7 +272,8 @@ fi
 rpm -q spacewalk-proxy-monitoring >/dev/null
 MONITORING=$?
 if [ $MONITORING -ne 0 ]; then
-        echo "You do not have monitoring installed. Do you want to install it?"
+        echo "You do not have monitoring installed."
+		echo "Do you want to install monitoring scout?"
 
         default_or_input "Will run '$YUM_OR_UPDATE spacewalk-proxy-monitoring'." INSTALL_MONITORING 'Y/n'
         INSTALL_MONITORING=$(yes_no $INSTALL_MONITORING)
@@ -200,7 +289,6 @@ else
         config_error 3 "Installation of package spacewalk-proxy-monitoring failed."
     fi
 fi
-ENABLE_SCOUT=0
 if [ $MONITORING -eq 0 ]; then
 	#here we configure monitoring
 	#and with cluster.ini
@@ -210,8 +298,11 @@ if [ $MONITORING -eq 0 ]; then
         default_or_input "Monitoring parent IP" MONITORING_PARENT_IP "$RESOLVED_IP"
         default_or_input "Enable monitoring scout" ENABLE_SCOUT "Y/n"
         ENABLE_SCOUT=$(yes_no $ENABLE_SCOUT)
-        MSG=$(echo -n "Your scout shared key (can be found on parent\nin $RHNCONF_DIR/cluster.ini as key scoutsharedkey)")
+        MSG=$(echo -n "Your scout shared key (can be found on parent
+in $RHNCONF_DIR/cluster.ini as key scoutsharedkey)")
         default_or_input "$MSG" SCOUT_SHARED_KEY ''
+else
+	ENABLE_SCOUT=0
 fi
 
 # size of squid disk cache will be 60% of free space on /var/spool/squid
@@ -233,12 +324,17 @@ sed -e "s|\${session.ca_chain:/usr/share/rhn/RHNS-CA-CERT}|$CA_CHAIN|g" \
 	    -e "s/\${session.use_ssl:0}/$USE_SSL/g" \
 	    -e "s/\${session.enable_monitoring_scout:0}/$ENABLE_SCOUT/g" \
         < $DIR/rhn.conf  > $RHNCONF_DIR/rhn.conf
-sed -e "s/\${session.enable_monitoring_scout:0}/$ENABLE_SCOUT/g" \
+if [ $MONITORING -eq 0 ]; then
+	sed -e "s/\${session.enable_monitoring_scout:0}/$ENABLE_SCOUT/g" \
             -e "s/\${session.rhn_monitoring_parent_ip}/$MONITORING_PARENT_IP/g" \
             -e "s/\${session.rhn_monitoring_parent}/$MONITORING_PARENT/g" \
             -e "s/\${session.scout_shared_key}/$SCOUT_SHARED_KEY/g" \
-        < cat $DIR/cluster.ini  > $RHNCONF_DIR/cluster.ini
+        < $DIR/cluster.ini  > $RHNCONF_DIR/cluster.ini
+fi
 
+# systemid need to be readable by apache/proxy
+chown root:apache $SYSCONFIG_DIR/systemid
+chmod 0640 $SYSCONFIG_DIR/systemid
 
 #Setup the cobbler stuff, needed to use koan through a proxy
 PROTO="http";
@@ -247,6 +343,12 @@ if [ $USE_SSL -eq 1 ]; then
 fi
 echo "ProxyPass /cobbler_api $PROTO://$RHN_PARENT/cobbler_api" > $HTTPDCONFD_DIR/cobbler-proxy.conf
 echo "ProxyPassReverse /cobbler_api $PROTO://$RHN_PARENT/cobbler_api" >> $HTTPDCONFD_DIR/cobbler-proxy.conf
+echo "ProxyPass /cblr $PROTO://$RHN_PARENT/cblr" >> $HTTPDCONFD_DIR/cobbler-proxy.conf
+echo "ProxyPassReverse /cblr $PROTO://$RHN_PARENT/cblr" >> $HTTPDCONFD_DIR/cobbler-proxy.conf
+echo "ProxyPass /cobbler $PROTO://$RHN_PARENT/cobbler" >> $HTTPDCONFD_DIR/cobbler-proxy.conf
+echo "ProxyPassReverse /cobbler $PROTO://$RHN_PARENT/cobbler" >> $HTTPDCONFD_DIR/cobbler-proxy.conf
+
+
 
 # lets do SSL stuff
 SSL_BUILD_DIR="/root/ssl-build"
@@ -284,7 +386,8 @@ if [ ! -f $SSL_BUILD_DIR/$RPM_CA ]; then
 	RPM_CA=$(grep noarch $SSL_BUILD_DIR/latest.txt)
 fi
 
-if [ ! -f $HTMLPUB_DIR/$RPM_CA ] || [ ! -f $HTMLPUB_DIR/RHN-ORG-TRUSTED-SSL-CERT ]; then
+if [ ! -f $HTMLPUB_DIR/$RPM_CA ] || [ ! -f $HTMLPUB_DIR/RHN-ORG-TRUSTED-SSL-CERT ] || \
+    ! diff $HTMLPUB_DIR/RHN-ORG-TRUSTED-SSL-CERT $SSL_BUILD_DIR/RHN-ORG-TRUSTED-SSL-CERT &>/dev/null; then
 	echo "Copying CA public certificate to $HTMLPUB_DIR for distribution to clients:"
 	cp $SSL_BUILD_DIR/RHN-ORG-TRUSTED-SSL-CERT $SSL_BUILD_DIR/$RPM_CA $HTMLPUB_DIR/
 fi
@@ -305,10 +408,12 @@ config_error $? "SSL key generation failed!"
 echo "Installing SSL certificate for Apache and Jabberd:"
 rpm -Uv $(/usr/bin/rhn-ssl-tool --gen-server --rpm-only --dir="$SSL_BUILD_DIR" 2>/dev/null |grep noarch.rpm)
 
-mv $HTTPDCONFD_DIR/ssl.conf $HTTPDCONFD_DIR/ssl.conf.bak
+if [ -e $HTTPDCONFD_DIR/ssl.conf ]; then
+	mv $HTTPDCONFD_DIR/ssl.conf $HTTPDCONFD_DIR/ssl.conf.bak
+fi
 sed -e "s|^SSLCertificateFile /etc/pki/tls/certs/localhost.crt$|SSLCertificateFile $HTTPDCONF_DIR/ssl.crt/server.crt|g" \
 	    -e "s|^SSLCertificateKeyFile /etc/pki/tls/private/localhost.key$|SSLCertificateKeyFile $HTTPDCONF_DIR/ssl.key/server.key|g" \
-	    -e "s|</VirtualHost>|SSLProxyEngine on\n</VirtualHost>|"
+	    -e "s|</VirtualHost>|SSLProxyEngine on\n</VirtualHost>|" \
         < $HTTPDCONFD_DIR/ssl.conf.bak  > $HTTPDCONFD_DIR/ssl.conf
